Extract cbindgen'd interface into its own crate
This means that we compile engine before we try running cbindgen over
it, so we get useful parse errors out of rustc, rather than errors
manifesting themselves as cbindgen going "There was a parse error
somewhere!"

I think ideally I would've moved the engine crate out into a new
subcrate, but this way around better preserves git history and avoids
merge conflicts.
