+++ src/rust/engine/engine_cffi/build.rs
@@ -30,15 +30,7 @@ use cbindgen;
 use cc;
 
 /*
-
-N.B. This build script is invoked by `cargo` by way of this configuration
-in our Cargo.toml:
-
-    [project]
-    ...
-    build = "src/cffi_build.rs"
-
-Within, we use the `gcc` crate to compile the CFFI C sources (`native_engine.c`)
+We use the `gcc` crate to compile the CFFI C sources (`native_engine.c`)
 generated by `bootstrap.sh` into a (private) static lib (`libnative_engine_ffi.a`),
 which then gets linked into the final `cargo build` product (the native engine binary).
 This process mixes the Python module initialization function and other symbols into the
@@ -100,8 +92,13 @@ fn main() -> Result<(), CffiBuildError> {
   let bindings_config_path = Path::new("cbindgen.toml");
   mark_for_change_detection(&bindings_config_path);
   mark_for_change_detection(Path::new("src"));
-  // Explicitly re-run if logging is modified because it's a hard-coded dep in cbindgen.toml.
-  mark_for_change_detection(Path::new("logging"));
+  // Explicitly re-run if engine or logging is modified because they're hard-coded deps in cbindgen.toml.
+  // Ideally we could just point at the engine crate here, but it's the workspace that contains all
+  // code, and we don't want to re-run on _any_ code changes (even though we'll probably re-run for
+  // them anyway), so we explicitly mark ../src and ../Cargo.toml for change-detection.
+  mark_for_change_detection(Path::new("../src"));
+  mark_for_change_detection(Path::new("../Cargo.toml"));
+  mark_for_change_detection(Path::new("../logging"));
 
   let scheduler_file_path = Path::new("src/cffi/scheduler.h");
   let crate_dir = env::var("CARGO_MANIFEST_DIR")?;
