+++ src/rust/engine/engine_cffi/cbindgen.toml
@@ -25,7 +25,15 @@ parse_deps = true
 # If more deps are added here, make sure to also add them to be rerun-triggers in src/cffi_build.rs.
 # We specify this explicitly because otherwise all transitive deps are parsed, and that takes a
 # very long time.
-include = ["logging"]
+#
+# Feel free to add more crates here if you need; basically any type that gets exposed over a C
+# interface needs to be listed here.
+#
+# We need to parse the engine crate to codegen PyResult (because externs are both something we use
+# from Python in engine, and something we expose to python in engine_cffi).
+#
+# We need to parse the logging crate to codegen PythonLogLevel and Destination.
+include = ["engine", "logging"]
 
 [export]
 # TODO: wrapped_PyInit_native_engine is declared twice if not excluded here, figure out why.
